--- a/openpathsampling/ensemble.py
+++ b/openpathsampling/ensemble.py
@@ -687,12 +687,15 @@ class EnsembleCombination(Ensemble):
         if Ensemble.use_shortcircuit:
             a = self.ensemble1(trajectory, trusted)
             logger.debug("Combination is " + self.__class__.__name__)
-            logger.debug("Combination: " + self.ensemble1.__class__.__name__ + 
-                         " is "+str(a))
-            logger.debug("Combination: " + self.ensemble2.__class__.__name__ +
-                         " is " +str(self.ensemble2(trajectory, trusted)))
-            logger.debug("Combination: returning " + 
-                         str(self.fnc(a,self.ensemble2(trajectory,trusted))))
+            logger.debug("Combination: " + self.ensemble1.__class__.__name__
+                         + " is "+str(a))
+            if logger.isEnabledFor(logging.DEBUG):
+                ens2 = self.ensemble2(trajectory, trusted)
+                logger.debug("Combination: " +
+                             self.ensemble2.__class__.__name__ +
+                             " is " +str(ens2))
+                logger.debug("Combination: returning " +
+                             str(self.fnc(a,ens2)))
             res_true = self.fnc(a, True)
             res_false = self.fnc(a, False)
             if res_false == res_true:
@@ -730,12 +733,14 @@ class EnsembleCombination(Ensemble):
             logger.debug("Combination is " + self.__class__.__name__)
             logger.debug("Combination.can_append: " + 
                          self.ensemble1.__class__.__name__ + " is "+str(a))
-            logger.debug("Combination.can_append: " 
-                         + self.ensemble2.__class__.__name__ +
-                         " is " +
-                         str(self.ensemble2.can_append(trajectory, trusted)))
-            logger.debug("Combination.can_append: returning " + 
-                         str(self.fnc(a,self.ensemble2.can_append(trajectory,trusted))))
+            if logger.isEnabledFor(logging.DEBUG):
+                ens2_can_append = self.ensemble2.can_append(trajectory,
+                                                            trusted)
+                logger.debug("Combination.can_append: " 
+                             + self.ensemble2.__class__.__name__ +
+                             " is " + str(ens2_can_append))
+                logger.debug("Combination.can_append: returning " + 
+                             str(self.fnc( a, ens2_can_append)))
             res_true = self._continue_fnc(a, True)
             res_false = self._continue_fnc(a, False)
             if res_false == res_true: