--- a/openpathsampling/volume.py
+++ b/openpathsampling/volume.py
@@ -118,7 +118,17 @@ class VolumeCombination(Volume):
         self.sfnc = str_fnc
 
     def __call__(self, snapshot):
-        return self.fnc(self.volume1.__call__(snapshot), self.volume2.__call__(snapshot))
+        # short circuit following JHP's implementation in ensemble.py
+        a = self.volume1(snapshot)
+        res_true = self.fnc(a, True)
+        res_false = self.fnc(a, False)
+        if res_false == res_true:
+            return res_true
+        else:
+            b = self.volume2(snapshot)
+            return self.fnc(a, b)
+        #return self.fnc(self.volume1.__call__(snapshot),
+                        #self.volume2.__call__(snapshot))
     
     def __str__(self):
         return '(' + self.sfnc.format(str(self.volume1), str(self.volume2)) + ')'