@dwhswenson 2016-03-28
Call graphs, generated by gprof2dot for `master` and for `ensemble_speed` branch. First, `master`:
![master](https://cloud.githubusercontent.com/assets/8178546/14080759/501a4a30-f507-11e5-942c-564ab6c85f7b.png)

Now `ensemble_speed`:
![ensemble_speed](https://cloud.githubusercontent.com/assets/8178546/14080783/633b8c1e-f507-11e5-855d-c1bbe77203d1.png)

Note the difference in the number of times `volume` calls `chain_dict`: it still seems a bit much at > 550k, but that's a big improvement over 12.3M.