@dwhswenson 2016-03-28
Passed tests. Ready for review.