@jhprinz 2016-03-28
Very nice! Really. This is how I like improvements... I think with all the changes in storage lately we should change the chaindict approach also a little. But that is different PR. Maybe cython will do some tricks. I think we might remove some convenience and trade it for speed. Especially the flexibility in storing and caching etc. is possible not necessary.

Looks good and will merge.